avances sem

+++ RowDaBoat-x64barebones-d4e1c147f975/Kernel/include/lib.h
@@ -12,6 +12,6 @@ uint64_t getRSP();
 
 void callTimerTick();
 
-uint64_t exchange(uint64_t * lock, uint64_t value);
+int exchange(uint8_t * lock, uint8_t value);
 
 #endif
